% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahg_estimate.R
\name{ahg_estimate}
\alias{ahg_estimate}
\title{Properly estimate AHG values}
\usage{
ahg_estimate(Q, Y = NULL, V = NULL, TW = NULL, allowance = 0.05,
  quiet = FALSE)
}
\arguments{
\item{Q}{streamflow timeseries}

\item{Y}{streamflow timeseries}

\item{V}{streamflow timeseries}

\item{TW}{streamflow timeseries}

\item{allowance}{streamflow timeseries}

\item{quiet}{streamflow timeseries}
}
\description{
Properly estimate AHG values
}
