% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ga2.R
\name{calc_ga_2}
\alias{calc_ga_2}
\title{Calculate GA AHG}
\usage{
calc_ga_2(Q, Y, V, TW, allowance = 0.05, r)
}
\arguments{
\item{Q}{streamflow timeseries}

\item{Y}{streamflow timeseries}

\item{V}{streamflow timeseries}

\item{TW}{streamflow timeseries}

\item{allowance}{streamflow timeseries}

\item{r}{streamflow timeseries}
}
\value{
data.frame
}
\description{
Calculate GA AHG
}
