% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ga.R
\name{calc_nsga}
\alias{calc_nsga}
\title{Calculate GA FHG}
\usage{
calc_nsga(df, allowance = 0.05, r, type = "nrmse")
}
\arguments{
\item{df}{hydraulic data.frame}

\item{allowance}{allowable deviation from continuity}

\item{r}{fit list}

\item{type}{metric to evaluate}
}
\value{
data.frame
}
\description{
Calculate GA FHG
}
\seealso{
Other FHG: 
\code{\link{best_optimal}()},
\code{\link{compute_fhg}()},
\code{\link{fhg_estimate}()},
\code{\link{mismash}()}
}
\concept{FHG}
