% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_fhg.R
\name{compute_ahg}
\alias{compute_ahg}
\title{Approximate AHG relationships}
\usage{
compute_ahg(Q, P, type = "relation")
}
\arguments{
\item{Q}{a stream flow time series}

\item{P}{a corresponding time series of a second hydraulic variable}

\item{type}{relationship being tested}
}
\value{
data.frame
}
\description{
Approximate AHG relationships using both OLS and NLS methods
}
\seealso{
Other AHG: 
\code{\link{ahg_estimate}()},
\code{\link{best_optimal}()},
\code{\link{calc_nsga}()},
\code{\link{min_max}()},
\code{\link{mismash}()}
}
\concept{AHG}
