% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_fhg.R
\name{compute_fhg}
\alias{compute_fhg}
\title{Approximate FHG relationships}
\usage{
compute_fhg(Q, P, type = NA)
}
\arguments{
\item{Q}{a stream flow time series}

\item{P}{a corresponding time series of a second hydraulic variable}

\item{type}{relationship being tested}
}
\value{
data.frame
}
\description{
Approximate FHG relationships using both OLS and NLS methods
}
