% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_shape.R
\name{compute_hydraulic_params}
\alias{compute_hydraulic_params}
\title{Approximate channel coeffiecent}
\usage{
compute_hydraulic_params(fit)
}
\arguments{
\item{fit}{output of ahg_estimate}
}
\value{
numeric
}
\description{
Approximate the hydraulic values from AHG fit
}
\seealso{
Other hydraulics: 
\code{\link{compute_channel_slope}()},
\code{\link{compute_n}()},
\code{\link{cross_section}()},
\code{\link{extract_thalweg}()},
\code{\link{slope_matrix}()}
}
\concept{hydraulics}
