% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_shape.R
\name{compute_n}
\alias{compute_n}
\title{Approximate Roughness}
\usage{
compute_n(df, S = 0.02)
}
\arguments{
\item{df}{a data.frame with at least Y and V.}

\item{S}{reach scale longitudinal slope (m/m). Default mean of the nhdplusV2}
}
\value{
numeric
}
\description{
Approximate median roughness using Manning Equation
}
\seealso{
Other hydraulics: 
\code{\link{compute_hydraulic_params}()},
\code{\link{cross_section}()}
}
\concept{hydraulics}
