% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{date_filter}
\alias{date_filter}
\title{Implements filtering by date}
\usage{
date_filter(df, years, keep_max = FALSE)
}
\arguments{
\item{df}{a data.frame with at least a date and Q field.}

\item{years}{the number of allowed history}

\item{keep_max}{Should the largest flow record be kept, even if older then "years"}
}
\value{
data.frame
}
\description{
Data is filtered when it is beyond a specified year threshold (e.g. 5 years old). The relative date is based on the newest 
observation in the data set. Optionally, the maximum flow (Q) record can be retained.
}
\seealso{
Other filters: 
\code{\link{mad_filter}()},
\code{\link{nls_filter}()},
\code{\link{qva_filter}()},
\code{\link{significance_check}()}
}
\concept{filters}
