% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopes.R
\name{extract_thalweg}
\alias{extract_thalweg}
\title{Extract Thalweg
From a data.frame of cross sections, a classified thalweg can be extracted as the connected LINESTRING}
\usage{
extract_thalweg(xs, crs = 5070)
}
\arguments{
\item{xs}{a data.frame containing cross sectional data. Required columns are hf_id, cs_id, X, Y, Z}

\item{crs}{the CRS of the XY coordinates}
}
\value{
XYZ LINESTRING object
}
\description{
Extract Thalweg
From a data.frame of cross sections, a classified thalweg can be extracted as the connected LINESTRING
}
\seealso{
Other hydraulics: 
\code{\link{compute_channel_slope}()},
\code{\link{compute_hydraulic_params}()},
\code{\link{compute_n}()},
\code{\link{cross_section}()},
\code{\link{slope_matrix}()}
}
\concept{hydraulics}
