% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhg_estimate.R
\name{fhg_estimate}
\alias{fhg_estimate}
\title{Properly estimate FHG values}
\usage{
fhg_estimate(df, allowance = 0.05, verbose = FALSE)
}
\arguments{
\item{df}{hydraulic data.frame}

\item{allowance}{allowed deviation from continuity}

\item{verbose}{should messages be emitted?}
}
\value{
list
}
\description{
Properly estimate FHG values
}
\details{
If filter_data is TRUE. An best fit curve is fit to the raw data relationships and is used to predict
A value for each Q. Data is only retained if the provided value is within +/- the `filter_threshold` (as a percent) of 
the estimated value
}
\seealso{
Other FHG: 
\code{\link{best_optimal}()},
\code{\link{calc_nsga}()},
\code{\link{compute_fhg}()},
\code{\link{mismash}()}
}
\concept{FHG}
