% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{mad_filter}
\alias{mad_filter}
\title{Implements filtering by median absolute deviation}
\usage{
mad_filter(df, envelope = 3)
}
\arguments{
\item{df}{a data.frame with at least a Q and one other AHG field (Y. TW, V).}

\item{envelope}{MAD envelope}
}
\value{
data.frame
}
\description{
An iterative outlier detection procedure is run based on to the linear regression residuals. 
Values of log-transformed TW, V, and Y residuals falling outside a specified median absolute deviation (MAD) envelope are excluded. 
Regression coefficients were recalculated and the outlier detection procedure was reapplied until no outliers are detected. This method
 was identified in \href{https://zenodo.org/record/7868764}{HyG}
}
\seealso{
Other filters: 
\code{\link{date_filter}()},
\code{\link{nls_filter}()},
\code{\link{qva_filter}()},
\code{\link{significance_check}()}
}
\concept{filters}
