% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mishmash.R
\name{mismash}
\alias{mismash}
\title{Compute all combos!}
\usage{
mismash(v, params, V, TW, Y, Q, r, allowancce)
}
\arguments{
\item{v}{}

\item{params}{}

\item{V}{}

\item{TW}{}

\item{Y}{}

\item{Q}{}
}
\value{
list
}
\description{
Compute all combos!
}
