% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ga.R
\name{nrmse}
\alias{nrmse}
\title{Normalized Root Mean Square Error}
\usage{
nrmse(sim, obs)
}
\arguments{
\item{sim}{numeric vector simulated values}

\item{obs}{numeric vector observed values}
}
\value{
numeric
}
\description{
Normalized root mean square error (NRMSE) between sim and obs, with treatment of missing values
}
\seealso{
Other evaluation: 
\code{\link{pbias}()}
}
\concept{evaluation}
