% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nwis}
\alias{nwis}
\title{Sample gage data
Manual measurements made at NWIS site 01096500}
\format{
#' @format ## `nwis`
A data frame with 245 rows and 6 columns:
\describe{
  \item{siteID}{NWIS ID}
  \item{date}{date of measurement}
  \item{Q}{Steamflow}
  \item{Y}{Depth}
  \item{V}{Velocity}
  \item{TW}{Top width}
}
}
\usage{
nwis
}
\description{
Sample gage data
Manual measurements made at NWIS site 01096500
}
\keyword{datasets}
