% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ga.R
\name{pbias}
\alias{pbias}
\title{Percent Bias}
\usage{
pbias(sim, obs)
}
\arguments{
\item{sim}{numeric vector simulated values}

\item{obs}{numeric vector observed values}
}
\value{
numeric
}
\description{
Percent Bias between sim and obs, 
with treatment of missing values.
}
\seealso{
Other evaluation: 
\code{\link{nrmse}()}
}
\concept{evaluation}
