% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{qva_filter}
\alias{qva_filter}
\title{Implements filtering by continuity}
\usage{
qva_filter(df, allowance = 0.05)
}
\arguments{
\item{df}{a data.frame with a Q, Y, TW, V and field.}

\item{allowance}{how much deviation from equality should be allowed (default = .05)}
}
\value{
data.frame
}
\description{
The function tests if the measured Q is outside of the expected range based on the product 
of measured velocity, top-width, and depth (e.g. Q≠vA)
}
\seealso{
Other filters: 
\code{\link{date_filter}()},
\code{\link{mad_filter}()},
\code{\link{nls_filter}()},
\code{\link{significance_check}()}
}
\concept{filters}
