% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{significance_check}
\alias{significance_check}
\title{Implements significance check}
\usage{
significance_check(df, pvalue = 0.05)
}
\arguments{
\item{df}{a data.frame with at least a Q and one other AHG field (Y. TW, V).}

\item{pvalue}{Significant p-value (default = .05)}
}
\value{
data.frame
}
\description{
The relationship between all supplied log transformed variables are computed.
If the p-value of any of these is less then the supplied p-value an error message is emitted.
}
\seealso{
Other filters: 
\code{\link{date_filter}()},
\code{\link{mad_filter}()},
\code{\link{nls_filter}()},
\code{\link{qva_filter}()}
}
\concept{filters}
