% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopes.R
\name{slope_matrix}
\alias{slope_matrix}
\alias{slope_weighted}
\alias{slope_mean}
\title{Calculate the gradient of line segments from a 3D matrix of coordinates}
\usage{
slope_matrix(mat, lonlat = TRUE)

slope_weighted(mat, lonlat = TRUE, directed = FALSE)

slope_mean(mat, lonlat = TRUE, directed = FALSE)
}
\arguments{
\item{mat}{Matrix containing coordinates and elevations.
The matrix should have three columns: X, Y, and Z. 
In data with geographic coordinates, Z values are assumed to be in
meters. In data with projected coordinates, Z values are assumed to have
the same units as the X and Y coordinates.}

\item{lonlat}{Are the elements provided in longitude/latitude coordinates?
By default, value is from the CRS of the routes (`sf::st_is_longlat(...)`).}
}
\value{
A vector of slopes associated with each LINE element
  The output value is a proportion representing the change in elevation
  for a given change in horizontal distance.
}
\description{
Calculate the gradient of line segments from a 3D matrix of coordinates
}
\seealso{
Other hydraulics: 
\code{\link{compute_channel_slope}()},
\code{\link{compute_hydraulic_params}()},
\code{\link{compute_n}()},
\code{\link{cross_section}()},
\code{\link{extract_thalweg}()}
}
\concept{hydraulics}
