% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{usgs_obs}
\alias{usgs_obs}
\title{Usgs Hydraulic Observation
Dataset containing USGS Hydraulic Observations}
\format{An object of class \code{list} of length 3544.}
\usage{
usgs_obs
}
\description{
Usgs Hydraulic Observation
Dataset containing USGS Hydraulic Observations
}
\keyword{datasets}
