% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhg_estimate.R
\name{ahg_estimate}
\alias{ahg_estimate}
\title{Properly estimate AHG values}
\usage{
ahg_estimate(
  df,
  allowance = 0.05,
  gen = 192,
  pop = 200,
  cprob = 0.4,
  mprob = 0.4,
  times = 1,
  scale = 1.5,
  full_fitting = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{df}{hydraulic data.frame with columns named (Q, V, TW, Y). 
Q and at least one other are required.}

\item{allowance}{allowed deviation from continuity}

\item{gen}{Number of generations to breed.}

\item{pop}{Size of population}

\item{cprob}{Crossover probability}

\item{mprob}{Mutation probability}

\item{times}{how many times (seeds) should nsga2 be run}

\item{scale}{should a scale factor be applied to data pre NSGA-2 fitting}

\item{full_fitting}{should all fits be returned?}

\item{verbose}{should messages be emitted?}
}
\value{
list
}
\description{
Properly estimate AHG values
}
\seealso{
Other AHG: 
\code{\link{best_optimal}()},
\code{\link{calc_nsga}()},
\code{\link{compute_ahg}()},
\code{\link{min_max}()},
\code{\link{mismash}()}
}
\concept{AHG}
