% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhg_estimate.R
\name{best_optimal}
\alias{best_optimal}
\title{Report best optimal}
\usage{
best_optimal(best, check, verbose = TRUE)
}
\arguments{
\item{best}{best performing method (character string)}

\item{check}{values to check against}

\item{verbose}{should messages be emitted}
}
\value{
vector
}
\description{
Report best optimal
}
\seealso{
Other FHG: 
\code{\link{calc_nsga}()},
\code{\link{compute_fhg}()},
\code{\link{fhg_estimate}()},
\code{\link{mismash}()}
}
\concept{FHG}
