% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhg_estimate.R
\name{best_optimal}
\alias{best_optimal}
\title{Report best optimal}
\usage{
best_optimal(best, check)
}
\arguments{
\item{best}{best performing method (character string)}

\item{check}{values to check against}
}
\value{
vector
}
\description{
Report best optimal
}
