% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ga.R
\name{calc_ga}
\alias{calc_ga}
\title{Calculate GA FHG}
\usage{
calc_ga(Q, Y, V, TW, allowance = 0.05, r, type = "nrmse")
}
\arguments{
\item{Q}{streamflow time series}

\item{Y}{river depth time series}

\item{V}{river velocity time series}

\item{TW}{river top width time series}

\item{allowance}{allowable deviation from continuity}

\item{r}{rrr}

\item{type}{metric to evaluate}
}
\value{
data.frame
}
\description{
Calculate GA FHG
}
