% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ga.R
\name{calc_nsga}
\alias{calc_nsga}
\title{Calculate NSGA2 AHG}
\usage{
calc_nsga(
  df,
  allowance = 0.05,
  r,
  scale = 2,
  gen = 96,
  pop = 500,
  cprob = 0.8,
  mprob = 0.05,
  times = 1
)
}
\arguments{
\item{df}{hydraulic data.frame}

\item{allowance}{allowable deviation from continuity}

\item{r}{fit list}

\item{scale}{should a scale factor be applied to data pre NSGA-2 fitting}

\item{gen}{Number of generations to breed.}

\item{pop}{Size of population}

\item{cprob}{Crossover probability}

\item{mprob}{Mutation probability}

\item{times}{how many times (seeds) should nsga2 be run}
}
\value{
data.frame
}
\description{
Calculate NSGA2 AHG
}
\seealso{
Other AHG: 
\code{\link{ahg_estimate}()},
\code{\link{best_optimal}()},
\code{\link{compute_ahg}()},
\code{\link{min_max}()},
\code{\link{mismash}()}
}
\concept{AHG}
