% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ahg.R
\name{compute_ahg}
\alias{compute_ahg}
\title{Approximate AHG realtionships}
\usage{
compute_ahg(Q, P, type = NA)
}
\arguments{
\item{Q}{a streamflow timeseries}

\item{P}{a corresponding timeseries of a second hydraulic variable}
}
\value{
data.frame
}
\description{
Approxiamte AHG realtionships using both OLS and NLS methods
}
