% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_shape.R
\name{cross_section}
\alias{cross_section}
\title{Approximate channel shape}
\usage{
cross_section(r, TW = 30, Ymax = 2, n = 30)
}
\arguments{
\item{r}{The corresponding Dingman's r coefficient}

\item{TW}{width of the channel at bankfull}

\item{Ymax}{maximum depth of the channel at bankfull}

\item{n}{the number of points to construct in the XS}
}
\value{
depth values every 1m along the cross section
}
\description{
Get a list of points from x axis of a cross section 
and max depth and produce depth values for 
those points based on channel shape
}
\seealso{
Other hydraulics: 
\code{\link{compute_hydraulic_params}()},
\code{\link{compute_n}()}
}
\concept{hydraulics}
