% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhg_estimate.R
\name{fhg_estimate}
\alias{fhg_estimate}
\title{Properly estimate FHG values}
\usage{
fhg_estimate(
  df,
  allowance = 0.05,
  gen = 192,
  pop = 200,
  cprob = 0.4,
  mprob = 0.4,
  times = 1,
  scale = 1.5,
  verbose = FALSE
)
}
\arguments{
\item{df}{hydraulic data.frame}

\item{allowance}{allowed deviation from continuity}

\item{gen}{Number of generations to breed.}

\item{pop}{Size of population}

\item{cprob}{Crossover probability}

\item{mprob}{Mutation probability}

\item{times}{how many times (seeds) should nsga2 be run}

\item{scale}{should a scale factor be applied to data pre NSGA-2 fitting}

\item{verbose}{should messages be emitted?}
}
\value{
list
}
\description{
Properly estimate FHG values
}
\seealso{
Other FHG: 
\code{\link{best_optimal}()},
\code{\link{calc_nsga}()},
\code{\link{compute_fhg}()},
\code{\link{mismash}()}
}
\concept{FHG}
