% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhg_estimate.R
\name{fhg_estimate}
\alias{fhg_estimate}
\title{Properly estimate FHG values}
\usage{
fhg_estimate(
  Q,
  Y = NULL,
  V = NULL,
  TW = NULL,
  allowance = 0.05,
  quiet = FALSE,
  forceGA = FALSE,
  nwis = NULL,
  location = NULL
)
}
\arguments{
\item{Q}{streamflow time series}

\item{Y}{depth time series}

\item{V}{velocity time series}

\item{TW}{top width time series}

\item{allowance}{allowed deviation from continuity}

\item{quiet}{should messages be emitted?}

\item{forceGA}{use GA over nsga2}

\item{nwis}{NWIS gage ID}

\item{location}{provide a locations}
}
\value{
list
}
\description{
Properly estimate FHG values
}
