% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mishmash.R
\name{mismash}
\alias{mismash}
\title{Compute all permutations of fitting types}
\usage{
mismash(v, V, TW, Y, Q, r, allowance)
}
\arguments{
\item{v}{values}

\item{V}{Velocity time series}

\item{TW}{Top width time series}

\item{Y}{Depth time series}

\item{Q}{Discharge time series}

\item{r}{list of fitting methods and parameters}

\item{allowance}{Allowable deviation from continuity}
}
\value{
list
}
\description{
Compute all permutations of fitting types
}
\seealso{
Other FHG: 
\code{\link{best_optimal}()},
\code{\link{calc_nsga}()},
\code{\link{compute_fhg}()},
\code{\link{fhg_estimate}()}
}
\concept{FHG}
